Make create tests use VC version 1.1 & 2.0.
